mmc: litex_mmc: fix cmd irq/completion sequencing
To ensure interrupts cannot be missed, we must (re)initialize
completions *before* setting up any transaction that will result
in the assertion of an interrupt. Waiting for completion should
follow *after* the hardware transaction setup, and before proceeding
with e.g. a polling loop acting as fall-back for when IRQ support
is unavailable.

Signed-off-by: Gabriel Somlo <gsomlo@gmail.com>